Rename file `db-helper.ts` -> `db.ts`
